requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
python-decouple==3.3
